Simplified NOP commands, and send ESC
NOP is used for heartbeat in normal mode, so just implement this.

If device doesn't answer to first NOP, try to turn it on by pressing ESC
through cbus bits. Fix needed as trying to turn on without success will
keep the algorithm looping and fail connection, leading to desync
between driver and Ekos.