Remote control through CCD interface and device buttons
With command to send buttons and frame read working, remote UI is now
functional, plan to use a guider CCD but not the usual GuiderInterface
controls unfortunately. Right now there's no need for navigation
scenarii.

Code is quite ugly, so now is time for a big cleanup.
Started reworking connection status into device state, but this is ugly.
Still a thread for connection keepalive, will revert to timer probably.