Fixed button repeat causing segfault.
As the remote buttons are a ATMOST1 type, there can happen that one of
them is released by the user when clicking before the message is handled
by the driver. This is especially visible on a remote server. In that
case looking for a "ON" switch returns null.

Also commented low-level byte exchanges in commands, and added a few
debug logs.