cmake: rewrite libusb-1.0 search
The current FindUSB-1.cmake is a mess: the search uses pkg-config only
as fallback, it mixes USB vs LIBUSB_1 prefixes, it mixes USB-1 vs USB_1
naming, it uses hardcoded paths, and the libusb_error_name() check is
done only on Linux.

As solution, write a brand-new FindUSB1.cmake module: it uses pkg-config
in the right way, it provides coherent naming for variables, and the
libusb_error_name() is done for any OS.

As side change, the define for USB1_HAS_LIBUSB_ERROR_NAME is not
automatically added to the build system, but written in a config-usb.h
file which is used where needed.