Removed the GPS slave properties
In functions `IsGetProperties` and `UpdatePropreties`, I removed the
`defineProperty(&GPSSlavingSP)` and `defineProperty(&GPSSlavingParamNP)`.

I am testing `indi-qhy` driver on QHY174GPS camera. With the firmware
released by QHY, if you try to get the GPS slave parameters the cooling
crash and does not work anymore. The problem can be resolved deleting the
`deineProperty`. I would like to prompt the QHY devs in order to go over
the firmware.