indi-svbony FIXED: Conflict of private variables in SVBONYCCD class
The private variable status in the SVBONYCCD class is shared among member functions, and the value of status may be changed asynchronously by member functions running in multi-threaded mode.
In practice, the use of the status variable is closed within each member function, and there is no need to share it as a member variable. Therefore, status was changed to a local variable within the member function that uses it.