Changed logic when PWM percent is zero
Previously, setting PWM to zero turned the port off. However, this prevented the status from being read so that if changed elsewhere, it would not be picked up.

Now, the ports are left on though with power of zero (same as turned off from a physical perspective).

Also, tidied up the return codes from setPortPWMDutyCyclePercent since this returns an integer rather than a bool,