AStarBox Plugin (#962)
* Initial Version of AStarBox

Initial version of AStarBox as a 3rd-party driver.

* New version of AStarBox.cpp

New version detects if there is a problem or not with the I2C connection.

Also, changed the names to be consistent with Rodolphe's files. (AStarBox vs CAStarBox), and changed astarbox to indi-astarbox.

* Update README

Update to improve format.

* Update README.md

Again change format.

* Another update

Formatting again

* And again

Another format change.

* Hopefully final update of README

I hope I have this right now!

* Readme

Getting name right in Readme

* Refactor of AStarBox.cpp

Called openAllPorts into connect so can use results to indicate whether connection successful  or not.

Moved getPortPWM for ports 5 and 6 into connect too - when in openAllPports m_bLinked was false at that point so no values collected.

Now called after m_bLlinked is true so that values are collected.

* Merged Rodolphe's updates and mine

Merged Rodolphe's updates and mine.

m_bPortsOpen is tested before opening all ports, but getting PWM5 and 6 is now in connect.

* Tidied up port numbers

Replaced 5 and 6 by their ENUMS instead for consistency.

* Changed logic when PWM percent is zero

Previously, setting PWM to zero turned the port off. However, this prevented the status from being read so that if changed elsewhere, it would not be picked up.

Now, the ports are left on though with power of zero (same as turned off from a physical perspective).

Also, tidied up the return codes from setPortPWMDutyCyclePercent since this returns an integer rather than a bool,