Restore touptek tail light and conversion gain setting on driver start-up (#1191)
* Restore touptek tail light setting in-camera

If tail light setting retrieved from the camera
differs from the one configured by the user, set the configured value
in the camera on driver start-up.

* Restore touptek conversion gain setting in-camera

If conversion gain setting retrieved from the camera
differs from the one configured by the user, set the configured value
in the camera on driver start-up.